Arrumando inconcistencias das frases no front
