+++ frontend/src/components/urls.js
@@ -66,4 +66,11 @@ function urlGen(microservice, endpoint) {
     return url;
 }
 
-export { urlGen, urlGenerator };
+function shuffle(array) {
+    for (let i = array.length - 1; i > 0; i -= 1) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [array[i], array[j]] = [array[j], array[i]];
+    }
+}
+
+export { urlGen, urlGenerator, shuffle };
